move group.list test
