+++ test/api/group/list.test.js
@@ -1,6 +1,21 @@
 const test = require('tape')
+const { Server } = require('../../helpers')
 
-test('group.list', { todo: true }, t => {
+test('list + get groups', (t) => {
+  const server = Server()
 
-  t.end()
+  server.tribes.create(null, (err, data) => {
+    t.error(err)
+    t.deepEqual(server.tribes.list(), [data.groupId])
+
+    const actualGroup = server.tribes.get(data.groupId)
+    const expectedGroup = {
+      key: data.groupKey,
+      root: data.groupInitMsg.key,
+      scheme: 'envelope-large-symmetric-group'
+    }
+    t.deepEqual(actualGroup, expectedGroup)
+
+    server.close(t.end)
+  })
 })
